requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.1
click==8.0.1
colorama==0.4.4
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.0
Pillow==8.3.1
pthflops==0.4.1
requests==2.26.0
torch==1.7.1+cpu
torchaudio==0.7.2
torchvision==0.8.2+cpu
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
